requirements.txt
Django~=5.1.1
djangorestframework~=3.15.2
django-cors-headers
djangorestframework-simplejwt
requests~=2.32.3
pandas~=2.2.2
numpy~=1.26.4
tensorflow~=2.17.0
scipy
stripe
africastalking~=1.2.8
xarray 
netCDF4~=1.7.1.post2
django-allauth
cryptography
scikit-learn~=1.5.2
rasterio~=1.3.11
pytest 
pytest-asyncio 
pytest-django
Pillow~=10.4.0
mysqlclient
pymongo[srv]~=4.9.1
google-auth
google-auth-oauthlib
pygal
highcharts-core
gunicorn
postgres
psycopg2-binary
psycopg2
openai
nltk~=3.9.1
scikit-learn
weasyprint~=62.3
matplotlib~=3.9.2
pytz~=2024.2
folium
sqlalchemy~=2.0.36
pymysql

earthengine-api~=1.1.5
protobuf~=4.25.4
django-geojson
django-leaflet  # Optional for front-end visualization
shapely
channels
channels-redis
fpdf
reportlab
